refactor: 패키지 오기명 변경 (user -> use)
